Desktop/comp90024_team_61-main/backend/fission/functions/scenario3/requirements.txt
flask
pandas
elasticsearch==8.10.1
python-dateutil
